Доработка по Спринту 6 по результатам review
