+++ src/controllers/InMemoryHistoryManager.java
@@ -53,15 +53,10 @@ public class InMemoryHistoryManager implements HistoryManager {
 
     @Override
     public void add(Task task) {
-        if (task != null) {
-            if (historyMap.containsKey(task.getId())) {
-                remove(task.getId());
-                historyMap.put(task.getId(), linkLast(task));
-            } else {
-                historyMap.put(task.getId(), linkLast(task));
-            }
-
-        }
+        if (task == null)
+            return;
+        remove(task.getId());
+        historyMap.put(task.getId(), linkLast(task));
     }
 
     @Override
@@ -80,6 +75,42 @@ public class InMemoryHistoryManager implements HistoryManager {
             node = node.getNext();
         }
         return tasksList;
+    }
+
+    private static class Node {
+        public Task data;
+        public Node next;
+        public Node prev;
+
+        public Node(Node prev, Task data, Node next) {
+            this.data = data;
+            this.next = next;
+            this.prev = prev;
+        }
+
+        public Task getData() {
+            return data;
+        }
+
+        public void setData(Task data) {
+            this.data = data;
+        }
+
+        public Node getPrev() {
+            return prev;
+        }
+
+        public void setPrev(Node prev) {
+            this.prev = prev;
+        }
+
+        public Node getNext() {
+            return next;
+        }
+
+        public void setNext(Node next) {
+            this.next = next;
+        }
     }
 }
