--- src/controllers/Node.java
@@ -1,41 +1,0 @@
-package controllers;
-
-import model.Task;
-
-public class Node {
-    public Task data;
-    public Node next;
-    public Node prev;
-
-    public Node(Node prev, Task data, Node next) {
-        this.data = data;
-        this.next = next;
-        this.prev = prev;
-    }
-
-    public Task getData() {
-        return data;
-    }
-
-    public void setData(Task data) {
-        this.data = data;
-    }
-
-    public Node getPrev() {
-        return prev;
-    }
-
-    public void setPrev(Node prev) {
-        this.prev = prev;
-    }
-
-    public Node getNext() {
-        return next;
-    }
-
-    public void setNext(Node next) {
-        this.next = next;
-    }
-}
-
-
